+++ tests/test_register.py
@@ -3,13 +3,15 @@ import xml.etree.ElementTree as ET
 import os
 import platform
 import dash
+import numpy as np
 import pytest
 import shutil
 import tifffile
 from rakaia.register.process import (
     dzi_tiles_from_image_path,
-    update_coregister_hash,
-    wsi_from_local_path)
+    update_wsi_hash,
+    wsi_from_local_path,
+    match_wsi_name_to_transformation_matrix)
 
 def test_parse_wsi_path(get_current_dir):
     files = wsi_from_local_path(os.path.join(get_current_dir, 'wsi'))
@@ -20,23 +22,26 @@ def test_parse_wsi_path(get_current_dir):
     assert not wsi_from_local_path(os.path.join(get_current_dir, 'steinbock'))
 
 def test_update_register_hash(get_current_dir):
-    new_hash = update_coregister_hash({}, os.path.join(get_current_dir, 'for_quant.tiff'))
-    assert 'for_quant.tiff' in new_hash.keys()
-    new_hash_2 = update_coregister_hash({}, os.path.join('not_dir', 'for_recolour.tiff'))
+    new_hash = update_wsi_hash({}, os.path.join(get_current_dir, 'for_quant.tiff'))
+    assert 'for_quant' in new_hash.keys()
+    new_hash_2 = update_wsi_hash({}, os.path.join('not_dir', 'for_recolour.tiff'))
     assert isinstance(new_hash_2, dash._callback.NoUpdate)
-    new_hash = update_coregister_hash(new_hash, os.path.join(get_current_dir, 'for_recolour.tiff'))
+    new_hash = update_wsi_hash(new_hash, os.path.join(get_current_dir, 'for_recolour.tiff'))
     assert len(new_hash) == 2
-    assert 'for_recolour.tiff' in new_hash.keys()
+    assert 'for_recolour' in new_hash.keys()
 
-    new_hash_list = update_coregister_hash({}, [os.path.join(get_current_dir, 'for_quant.tiff'),
-                                      os.path.join(get_current_dir, 'for_recolour.tiff')])
+    new_hash_list = update_wsi_hash({}, [os.path.join(get_current_dir, 'for_quant.tiff'),
+                                         os.path.join(get_current_dir, 'for_recolour.tiff')])
     assert len(new_hash_list) == 2
-    assert 'for_recolour.tiff' in new_hash_list.keys()
+    assert 'for_recolour' in new_hash_list.keys()
 
-    assert isinstance(update_coregister_hash({}, None), dash._callback.NoUpdate)
+    assert isinstance(update_wsi_hash({}, None), dash._callback.NoUpdate)
 
-
-
+def test_import_wsi_transform(get_current_dir):
+    transform_hash = update_wsi_hash({}, os.path.join(get_current_dir, 'affine_transform.csv'), True)
+    assert 'affine_transform' in transform_hash
+    assert isinstance(transform_hash['affine_transform'], np.ndarray)
+    assert transform_hash['affine_transform'].shape == (3, 3)
 
 @pytest.mark.skipif(platform.system() != 'Linux',
                     reason='install pyvips only for Linux during testing')
@@ -79,3 +84,12 @@ def test_generate_dzi_tiles(get_current_dir):
                                   download_dir)
 
         assert not os.path.isfile(os.path.join(download_dir, 'coregister.dzi'))
+
+
+def test_name_match_wsi_transform():
+    wsi = "Lung_V1_FFPE"
+    transform_options = ['Xenium_Skin_FFPE_V1', 'Lung_v2', 'Pancreas_FFPE', 'Lung_V1_FFPEalign']
+    assert match_wsi_name_to_transformation_matrix(wsi, transform_options)
+    assert not match_wsi_name_to_transformation_matrix(wsi, list(transform_options[-1]))
+    assert not match_wsi_name_to_transformation_matrix(wsi, None)
+    assert not match_wsi_name_to_transformation_matrix('Lung_V1', ['Lung_v2', 'Lung_FFPEalign'])